Use the correct type for GLintptr
On 32bit, intptr_t has a different type than khronos_intptr_t.

https://bugzilla.gnome.org/show_bug.cgi?id=691296
(cherry picked from commit a77222385812a696a999a871cde439f50b648f69)

+++ cogl/cogl-gles2-types.h
@@ -35,7 +35,7 @@ typedef float            GLclampf;
 typedef int32_t          GLfixed;
 
 /* GL types for handling large vertex buffer objects */
-typedef intptr_t         GLintptr;
+typedef signed long int  GLintptr;
 typedef long             GLsizeiptr;
 
 /* OpenGL ES core versions */
